Report refactoring (#11)
• Functions of the 'report.utils' module have been integrated into the 'Report' class.
• Metaclass 'MetaReport' has been dropped.
• Functions of the 'report.composers' module have been integrated into the 'Report' class.
• The 'Report' class has been moved into the new 'linnote.report' module. The 'linnote.report' package has been dropped.
