+++ linnote/templates/ranking.html
@@ -43,7 +43,7 @@
                 </table>
 
                 <svg width="100%" height="50%" preserveAspectRatio="X432Y288" viewBox="0 0 432 288">
-                    {{ data.graph }}
+                    {{ data.histogram }}
                 </svg>
 
                 <table id="grade">
